Make the candidate list of event files much larger
We had an issue where the touchscreen device was event11, and we were
only checking event files 0-10. Now we check 0-99 (inclusive), and that
should be sufficient in all cases.
If not, we have to do something more complicated to get the correct
event stream. Maybe look at how evtest does it?